+++ main.py
@@ -2,34 +2,51 @@
 ---------------------------------------------------------
 Projeto : Caixa Express
 Arquivo : main.py
+Autor   : Dylan Ryan Pereira Santos
+Versão  : 0.1.0
+---------------------------------------------------------
+Descrição:
+Ponto de entrada principal do Caixa Express.
 ---------------------------------------------------------
 """
-from ui.app import CaixaExpressApp
-from config.constants import CONFIG_FILE
-from config.constants import LOG_DIR
 
-from config.version import AppInfo
-
+from config.constants import CONFIG_FILE, LOG_DIR
 from core.config_manager import ConfigManager
 from core.database import DatabaseManager
 from core.logger_manager import LoggerManager
+from ui.app import CaixaExpressApp
 
 
 def main() -> None:
+    """Inicializa os componentes principais da aplicação."""
 
     logger = LoggerManager(LOG_DIR)
 
-    logger.info("Inicializando Caixa Express.")
+    logger.info(
+        "Inicializando Caixa Express."
+    )
 
-    ConfigManager(CONFIG_FILE)
+    config_manager = ConfigManager(
+        CONFIG_FILE
+    )
 
-    DatabaseManager(logger)
+    database = DatabaseManager(
+        logger
+    )
 
-    app = CaixaExpressApp()
+    try:
+        app = CaixaExpressApp(
+            config_manager
+        )
 
-    app.run()
+        app.run()
 
-    logger.info("Sistema encerrado.")
+    finally:
+        database.close()
+
+        logger.info(
+            "Sistema encerrado."
+        )
 
 
 if __name__ == "__main__":